Introduce the BuddyPress_Skeleton_Screens class to manage screens
This class is first filtering bp_located_template to check if a template was found in active theme. Then if the theme is using BP Theme Compat, it will return false so that the Theme compat feature
can manage the example component screens templates.
The filter applyed to bp_get_template_stack will add to the templates stack the template dir of the example component.
Then when the screen_one screen callback will be run when displaying this screen, a filter will be applyed to bp_get_template_part in order to load screen-one template instead of members/single/plugins one

We also need to replace the screen one screen callback in the component loader to use the one include in this new class.
Finally removing the function bp_example_load_template_filter() which will no more be used.