Move all that relates to notifications in bp-example-notifications.php
In the component loader, include the bp-example-notifications.php file only the the notifications component is active.
Use the bp_example_send_high_five hook to run the screen and email notifications function.
Check if settings component is active before adding the link to email notifications settings in the bottom of the mail sent