Move all that relates to activity in bp-example-activity.php file
Only include this file if the activity component is active from the component loader.
Use hooks to record the activities.
Register activity actions into the activity actions global for the component in order to be in WP-Admin dropdown type and in front end activity directory and member profile filters. Register the callback format functions to benefit from the improvements introduced in BuddyPress 2.0 about translatable strings in activity actions.